tests

+++ src/test/bash/files/equals_test.sh
@@ -172,7 +172,21 @@ ${ASSERTS_TEXT}
 done
 rm "${TMP_PATH}"
 
-echo 'Not implemented!'; exit 1 # todo
+:> "${STDOUT}"
+:> "${STDERR}"
+TMP_PATH="$(mktemp)"
+ASSERTS_TEXTS=('42' 'foobar' 'barfoo' ' foo' 'foo ' $'\nfoo' $'\tfoo' $'foo\n' $'foo\t')
+for ASSERTS_TEXT in "${ASSERTS_TEXTS[@]}"; do
+ printf '%s' "${ASSERTS_TEXT}" > "${TMP_PATH}"
+ "${SCRIPT}" "${TMP_PATH}" "${ASSERTS_TEXT}" >"${STDOUT}" 2>"${STDERR}"; CODE=$?
+ if [[ "${CODE}" != '0' ]]; then
+  echo "Code(${CODE}) error!" >&2; exit 1; fi
+ [[ -s "${STDOUT}" ]] && exit 1
+ ACTUAL_VALUE="$(<"${STDERR}")"
+ if [[ -n "${ACTUAL_VALUE}" ]]; then
+  echo "Actual value(${#ACTUAL_VALUE}) is: \"${ACTUAL_VALUE}\"!" >&2; exit 1; fi
+done
+rm "${TMP_PATH}"
 
 rm "${STDOUT}"
 rm "${STDERR}"
